// frontend/src/openapi/compassClient/models/CategoryDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { UserDto } from './UserDto';
import {
    UserDtoFromJSON,
    UserDtoFromJSONTyped,
    UserDtoToJSON,
} from './UserDto';

/**
 * 
 * @export
 * @interface CategoryDto
 */
export interface CategoryDto {
    /**
     * 
     * @type {number}
     * @memberof CategoryDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof CategoryDto
     */
    name?: string;
    /**
     * 
     * @type {number}
     * @memberof CategoryDto
     */
    minimumValue?: number;
    /**
     * 
     * @type {number}
     * @memberof CategoryDto
     */
    maximumValue?: number;
    /**
     * 
     * @type {Array<UserDto>}
     * @memberof CategoryDto
     */
    categoryOwners?: Array<UserDto>;
}

/**
 * Check if a given object implements the CategoryDto interface.
 */
export function instanceOfCategoryDto(value: object): boolean {
    return true;
}

export function CategoryDtoFromJSON(json: any): CategoryDto {
    return CategoryDtoFromJSONTyped(json, false);
}

export function CategoryDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): CategoryDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
        'minimumValue': json['minimumValue'] == null ? undefined : json['minimumValue'],
        'maximumValue': json['maximumValue'] == null ? undefined : json['maximumValue'],
        'categoryOwners': json['categoryOwners'] == null ? undefined : ((json['categoryOwners'] as Array<any>).map(UserDtoFromJSON)),
    };
}

export function CategoryDtoToJSON(value?: CategoryDto | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'id': value['id'],
        'name': value['name'],
        'minimumValue': value['minimumValue'],
        'maximumValue': value['maximumValue'],
        'categoryOwners': value['categoryOwners'] == null ? undefined : ((value['categoryOwners'] as Array<any>).map(UserDtoToJSON)),
    };
}